+++ packages/br-validators/src/feriados/constants.ts
@@ -1,0 +1,2 @@
+export const FERIADOS_FIXED_COUNT = 9;
+export const FERIADOS_MOVABLE_COUNT = 3;